Introduce rootfs UUID and use it as GPT partition UUID on install.
The rationale behind this commit is to make easier for grub to
find the partition UUID is booting from. A UUID is generated on
creation of the rootfs image, and such UUID is passed to the
generated grub configuration file.

The generation of boot configuration file is also generalised,
to allow other bootloader (e.g., ELF native bootloaders) to
be configured automatically.

On install, after copying the image the installer sets the
rootfs UUID as the GPT partition UUID.